requirements.txt
streamlit>=1.28.0
flask>=2.3.0
pypdf>=3.17.0
pdfminer.six>=20221105
tabula-py>=2.8.0
pdfquery>=0.4.3
PyMuPDF>=1.23.0
pdfplumber>=0.10.0
llamaparse>=0.3.0
pandas>=2.0.0
python-docx>=0.8.11
openpyxl>=3.1.0
beautifulsoup4>=4.12.0
markdown>=3.5.0
lxml>=4.9.0
Pillow>=10.0.0
pytesseract>=0.3.10
requests>=2.31.0